No commit message

+++ libs/utils/menu/__init__.py
@@ -422,33 +422,34 @@ class Menu(Generic[T]):
         self.on_close()
 
     def voice_input(self):
-        from ai.openai.speech_to_text import convert_audio_to_text
-
-        from utils.menu.asynctaskmenu import AsyncTaskMenu
-        from utils.menu.recordmenu import RecordMenu
-
-        record_menu = RecordMenu()
-        record_menu.exec()
-        out_file = record_menu.get_output_file()
-        if not out_file:
-            return
-
-        stt_menu = AsyncTaskMenu(
-            lambda: convert_audio_to_text(file=out_file),
-            prompt="(Converting audio to text...)",
-        )
         try:
+            from ai.openai.speech_to_text import convert_audio_to_text
+
+            from utils.menu.asynctaskmenu import AsyncTaskMenu
+            from utils.menu.recordmenu import RecordMenu
+
+            record_menu = RecordMenu()
+            record_menu.exec()
+            out_file = record_menu.get_output_file()
+            if not out_file:
+                return
+
+            stt_menu = AsyncTaskMenu(
+                lambda: convert_audio_to_text(file=out_file),
+                prompt="(Converting audio to text...)",
+            )
             stt_menu.exec()
+
+            if os.path.exists(out_file):
+                os.remove(out_file)
+
+            text = stt_menu.get_result()
+            if text:
+                self.insert_text(text)
+                if not record_menu.space_pressed:
+                    self.on_enter_pressed()
         except Exception as e:
             self.set_message(f"ERROR: {e}")
-            return
-
-        os.remove(out_file)
-        text = stt_menu.get_result()
-        if text:
-            self.insert_text(text)
-            if not record_menu.space_pressed:
-                self.on_enter_pressed()
 
     def __select_all(self):
         self.set_selection(0, -1)
@@ -620,6 +621,7 @@ class Menu(Generic[T]):
         stdscr.nodelay(False)
         stdscr.timeout(1000)
         Menu._stdscr = stdscr
+        Menu._should_update_screen = True
 
     @staticmethod
     def _get_color_pair(fg: Union[int, str], bg: Union[int, str] = -1) -> int:
